+++ libyul/AsmParser.cpp
@@ -755,7 +755,13 @@ YulName Parser::expectAsmIdentifier()
 {
 	YulName name{currentLiteral()};
 	if (currentToken() == Token::Identifier && m_dialect.findBuiltin(name.str()))
-		fatalParserError(5568_error, "Cannot use builtin function name \"" + name.str() + "\" as identifier name.");
+		// Non-fatal. We'll continue and wrongly parse it as an identifier. May lead to some spurious
+		// errors after this point, but likely also much more useful ones.
+		m_errorReporter.parserError(
+			5568_error,
+			currentLocation(),
+			"Cannot use builtin function name \"" + name.str() + "\" as identifier name."
+		);
 	// NOTE: We keep the expectation here to ensure the correct source location for the error above.
 	expectToken(Token::Identifier);
 	return name;